+++ internal/handlers/auth.go
@@ -89,41 +89,17 @@ func (h *Handler) Register(c *gin.Context) {
 	})
 }
 
-func (h *Handler) GetUser(c *gin.Context) {
+func (h *Handler) Logout(c *gin.Context) {
 	userID, exists := c.Get("user_id")
 	if !exists {
 		c.JSON(http.StatusUnauthorized, gin.H{"error": "User not authenticated"})
 		return
 	}
 
-	user, err := h.Service.GetUser(userID.(string))
-	if err != nil {
-		c.JSON(http.StatusNotFound, gin.H{"error": "User not found"})
-		return
-	}
-
-	c.JSON(http.StatusOK, user)
-}
-
-func (h *Handler) DeleteUser(c *gin.Context) {
-	userID := c.Param("userId")
-	currentUserID, exists := c.Get("user_id")
-
-	if !exists {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "User not authenticated"})
-		return
-	}
-
-	// Users can only delete their own account
-	if userID != currentUserID.(string) {
-		c.JSON(http.StatusForbidden, gin.H{"error": "Cannot delete other users"})
-		return
-	}
-
-	if err := h.Service.DeleteUser(userID); err != nil {
+	if err := h.Service.Logout(userID.(string)); err != nil {
 		c.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
 	}
 
-	c.JSON(http.StatusOK, gin.H{"message": "User deleted successfully"})
+	c.JSON(http.StatusOK, gin.H{"message": "User logged out successfully"})
 }
