+++ internal/services/auth.go
@@ -45,78 +45,10 @@ func (s *Service) Login(email, password string) (models.User, error) {
 	return user, nil
 }
 
-func (s *Service) GetUser(userID string) (models.User, error) {
-	userDB, err := s.Storage.GetUserByID(userID)
-	if err != nil {
-		return models.User{}, err
-	}
-
-	user := mapToUser(userDB)
-
-	return user, nil
-}
-
-func (s *Service) DeleteUser(userID string) error {
-	// Delete user resources first
-	if err := s.DeleteUserResources(userID); err != nil {
-		return errors.Wrap(err, "failed to delete user resources")
-	}
-
-	// Finally, delete the user itself
-	if err := s.Storage.DeleteUserByID(userID); err != nil {
-		return errors.Wrap(err, "failed to delete user")
+func (s *Service) Logout(userID string) error {
+	if err := s.Storage.RevokeToken(userID); err != nil {
+		return errors.Wrap(err, "failed to logout user")
 	}
 
 	return nil
-}
-
-func (s *Service) DeleteUserResources(userID string) error {
-	// Get sites owned by the user
-	userSites, err := s.Storage.GetSitesByUserID(userID)
-	if err != nil {
-		return errors.Wrap(err, "failed to get user sites")
-	}
-	var siteIDs []string
-	for _, site := range userSites {
-		siteIDs = append(siteIDs, site.ID)
-	}
-
-	// Get collections from the sites
-	siteCollections, err := s.Storage.GetCollectionsFromSitesIDs(siteIDs)
-	if err != nil {
-		return errors.Wrap(err, "failed to get collections from sites")
-	}
-	var collectionIDs []string
-	for _, collection := range siteCollections {
-		collectionIDs = append(collectionIDs, collection.ID)
-	}
-
-	// Delete entries by collection IDs
-	if len(collectionIDs) > 0 {
-		if err := s.Storage.DeleteEntriesByCollectionIDs(collectionIDs); err != nil {
-			return errors.Wrap(err, "failed to delete entries")
-		}
-	}
-
-	// Delete collections by site IDs
-	if len(siteIDs) > 0 {
-		if err := s.Storage.DeleteCollectionsBySiteIDs(siteIDs); err != nil {
-			return errors.Wrap(err, "failed to delete collections")
-		}
-	}
-
-	// Delete sites by user ID
-	if err := s.Storage.DeleteSitesByUserID(userID); err != nil {
-		return errors.Wrap(err, "failed to delete sites")
-	}
-
-	return nil
-}
-
-func mapToUser(userDB storage.UserDB) models.User {
-	return models.User{
-		ID:       userDB.ID,
-		Email:    userDB.Email,
-		Username: userDB.Username,
-	}
 }
