+++ internal/storage/user.go
@@ -1,5 +1,7 @@
 package storage
 
+import "github.com/pkg/errors"
+
 func (s *Storage) CreateUser(user UserDB) error {
 	created := s.DB.Create(&user)
 	if created.Error != nil {
@@ -24,6 +26,13 @@ func (s *Storage) GetUserByID(id string) (UserDB, error) {
 		return user, err
 	}
 	return user, nil
+}
+
+func (s *Storage) RevokeToken(userID string) error {
+	if err := s.DB.Model(&UserDB{}).Where("id = ?", userID).Update("token", "").Error; err != nil {
+		return errors.Wrap(err, "failed to revoke token for user")
+	}
+	return nil
 }
 
 func (s *Storage) DeleteUserByID(id string) error {
